pyproject.toml
[tool.poetry]
name = "project-async-python-00"
version = "0.1.0"
description = "An aysnc python project to implement leaky bucket algorithm in message queues."
authors = ["shweta-0812 <shweta.itenggie@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "=3.12.3"
fastapi = {extras = ["standard"], version = "=0.112.2"}
pydantic = "=2.8.2"
flake8 = "=7.1.1"
uvicorn = "^0.30.6"
asyncio = "=3.4.3"
python-dotenv = "=1.0.1"
setuptools = "=74.1.2"
redis = "=5.0.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
